Renamed 'light.h' and 'light.cpp' to 'Light.h' and 'Light.cpp'
