+++ Engine/Core/Lights/Light.cpp
@@ -1,4 +1,4 @@
-#include "light.h"
+#include "Light.h"
 
 Light::Light(glm::vec3 position, glm::vec3 colour, float strength) {
 	this->position = position;
